Move lookup table construction to `_lookup.py`
